+++ pkg/breaker/breaker_test.go
@@ -1,4 +1,4 @@
-package d4f
+package breaker
 
 import (
 	"context"
